♻️ refactor: modularize event listener logic for better structure
- Moved addEventListener logic from main.js to respective component files
- Introduced init functions in each component and call them from main.js
